+++ app/scripts/directives/graph-directives.js
@@ -118,7 +118,9 @@ angular.module('lizard-nxt')
    * @description       Stores the graph directives data and update functions
    */
   graphCtrl = function ($scope, Graph) {
+
     this.setData = function (scope) {
+
       // Provide defaults for backwards compatability
       this.data = scope.data || [];
       this.keys = scope.keys || {x: 0, y: 1};
@@ -127,6 +129,7 @@ angular.module('lizard-nxt')
         y: scope.ylabel || ''
       };
     };
+
     this.setData($scope);
 
     this.graph = {};